requirements.txt
celery==5.3.4
Django==5.2.1
stripe==7.8.0
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple==3.8
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
djangorestframework==3.16.0